Add nginx for SSL termination
haproxy is designed as single process/single threaded which becomes a
problem if used for SSL termination. There is nbproc to start multiple
processes but it has it's downsides and even the maintainer advise
against. I looked at stud for SSL termination but that project seems
pretty much dead, so I went for nginx.