Fix run script not properly killing ucarp
This is a interesting issue which took me quite some time to debug:
set -e makes the bash script exit on error, since a died/killed process
is an error, the script gets aborted and the shutdown function never
gets executed. But since there are still processes running in the
background, it won't exit either.